move everything to the programs directory
