+++ core/src/programs/PeopleMemory/PersonModel.ts
